requirements.txt
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.2
pyyaml==6.0.2